+++ bin/start.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+node --harmony bin/service.js "$@"